Refactor ~erc-social-graph-draw~
I was having some major issues getting this function to properly find
the buffer that it was using and writing the hash out to it. Basically
the '(unless file' clause was causing it to not run at all, so I
refactored this to precalc the file in the initial let binding and work
with it via find-file.